+++ src/main/java/com/whattheburger/backend/domain/cart/CartCalculator.java
@@ -9,6 +9,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -18,10 +19,10 @@ import java.util.stream.Collectors;
 public class CartCalculator {
     private final ProductCalculator productCalculator = new ProductCalculator();
 
-    public Double calculate(CalculatorDto calculatorDto) {
+    public BigDecimal calculate(CalculatorDto calculatorDto) {
         List<ProductDetail> productDetails = calculatorDto.getProductDetails();
         log.info("product count: {}", productDetails.size());
-        Double totalPrice = productCalculator.calculateTotalPrice(productDetails);
+        BigDecimal totalPrice = productCalculator.calculateTotalPrice(productDetails);
         return totalPrice;
     }
 }
